Split out ValidationError into its own mod
